# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/get_catalogue_offers": {
      "get": {
        "tags": ["geo"],
        "summary": "Get Offers",
        "operationId": "get_offers_get_catalogue_offers_get",
        "parameters": [
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "title": "Longitude" }
          },
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "title": "Latitude" }
          },
          {
            "name": "DeprecationDays",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Deprecationdays"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_new_market": {
      "post": {
        "tags": ["geo"],
        "summary": "Post Market",
        "operationId": "post_market_add_new_market_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddSupermarketRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/post_catalogue": {
      "post": {
        "tags": ["geo"],
        "summary": "Post Catalogue",
        "operationId": "post_catalogue_post_catalogue_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_catalogue_post_catalogue_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/get_supermarkets_close": {
      "get": {
        "tags": ["geo"],
        "summary": "Get Supermarkets Close",
        "operationId": "get_supermarkets_close_get_supermarkets_close_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "number" }, { "type": "null" }],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "number" }, { "type": "null" }],
              "title": "Lon"
            }
          },
          {
            "name": "radius_meters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 100,
              "default": 2000,
              "title": "Radius Meters"
            }
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "array", "items": { "type": "string" } },
                { "type": "null" }
              ],
              "title": "Chains"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddSupermarketRequest": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "address": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address"
          },
          "city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "City"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category"
          },
          "postcode": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Postcode"
          },
          "latitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Longitude"
          }
        },
        "type": "object",
        "title": "AddSupermarketRequest"
      },
...
      "Body_post_catalogue_post_catalogue_post": {
        "properties": {
          "catalogue": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Catalogue"
          },
          "latitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Longitude"
          },
          "lat": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Lon"
          },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "address": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address"
          },
          "categories": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Categories"
          },
          "city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "City"
          },
          "postcode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Postcode"
          }
        },
        "type": "object",
        "required": ["catalogue"],
        "title": "Body_post_catalogue_post_catalogue_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}